pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ldsrs2-launcher"
version = "0.1.0"
description = "Production-grade Slurm launcher for LDSR-S2 inference"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Open source (MIT)"}
authors = [
  {name = "Davide Mattioli"}
]
dependencies = [
  "PyYAML>=6.0",
  "numpy>=1.24",
  "rasterio>=1.3",
  "rioxarray>=0.15",
  "pyproj>=3.6",
  "omegaconf>=2.3",
  "torch>=2.1",
  "cubo>=0.3",
  "opensr-model>=0.1",
  "opensr-utils>=0.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
ldsrs2 = "ldsrs2_launcher.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.pyright]
include = ["src", "tests"]
extraPaths = ["src"]
